fix(test): fix type mismatch in query

+++ src/main/java/dev/genesshoan/fitnesstrackerapi/routine/dto/RoutineListItemDTO.java
@@ -10,6 +10,6 @@ public record RoutineListItemDTO(
         @Schema(description = "The routine's unique identifier", example = "123e4567-e89b-12d3-a456-426614174000")
                 UUID id,
         @Schema(description = "The routine's name", example = "Push Day") String name,
-        @Schema(description = "The number of exercises in the routine", example = "5") int exerciseCount,
+        @Schema(description = "The number of exercises in the routine", example = "5") Long exerciseCount,
         @Schema(description = "The date and time the routine was last updated", example = "2024-01-01T12:00:00")
                 LocalDateTime updatedAt) {}
